refactor: change naming to kebub-case
