+++ src/__tests__/home.test.tsx
@@ -1,4 +1,4 @@
-import Home from '@/app/page';
+import Home from '@/app/home';
 import { render, screen } from '@testing-library/react';
 import { expect, test } from 'vitest';
