+++ src/components/footer-component.tsx
